DeploymentCode/lambdas/calculate_availability_v2/requirements.txt
httpx
PyJWT
cryptography